+++ CHANGELOG.md
@@ -5,7 +5,7 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.6.0] - 2022-11-28
 
 ### Added
 
@@ -16,7 +16,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 * `ResourceUsage` and functions related with getting resource usage.
 
-## [0.5.0]
+## [0.5.0] - 2022-11-07
 
 ### Changed
 
@@ -138,7 +138,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 * `send_frame` and `recv_frame` to send and receive length-delimited frames.
 
-[Unreleased]: https://github.com/petabi/oinq/compare/0.5.0...main
+[0.6.0]: https://github.com/petabi/oinq/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/petabi/oinq/compare/0.4.1...0.5.0
 [0.4.1]: https://github.com/petabi/oinq/compare/0.4.0...0.4.1
 [0.4.0]: https://github.com/petabi/oinq/compare/0.3.0...0.4.0